requirements/requirements.txt
requests
beautifulsoup4
pandas
numpy
ipykernel
plotly
nbformat>=5.0.0
boto3
sqlalchemy
pymysql
python-dotenv